Move tests into another file
